+++ backend/templates/proposal_template_cerf.json
@@ -208,16 +208,7 @@
           {
             "name": "Categories",
             "format_type": "text",
-            "instructions": """ Use one of the following values: [
-              "Displacement related",
-              "IDP",
-              "Refugees",
-              "Returnees",
-              "Host Communities",
-              "Non-Displacement related",
-              "Total Directly Targeted",
-              "Persons with Disabilities"
-            ] """
+            "instructions": "Use one of the following values: ["Displacement related", "IDP", "Refugees", "Returnees","Host Communities", "Non-Displacement related", "Total Directly Targeted",  "Persons with Disabilities"] "
           },
           {
             "name": "Women (>18)",
@@ -257,15 +248,7 @@
           {
             "name": "Categories",
             "format_type": "text",
-            "instructions": """ Use one of the following values: [
-              "Displacement related",
-              "IDP",
-              "Refugees",
-              "Returnees",
-              "Host Communities",
-              "Non-Displacement related",
-              "Total Directly Targeted"
-            ]"""
+            "instructions": "Use one of the following values: ["Displacement related", "IDP", "Refugees", "Returnees","Host Communities", "Non-Displacement related", "Total Directly Targeted",  "Persons with Disabilities"] "
           },
           {
             "name": "Women (>18)",
