+++ playwright/tests/test_2_create_proposal.py
@@ -1,95 +1,152 @@
 import re
 import pytest
-from playwright.sync_api import Page, expect
+import os
+from playwright.sync_api import Playwright, sync_playwright, Page, expect
 
-def test_generate_new_proposal(page: Page):
+def test_generate_new_proposal():
     """
-    Tests that a user can generate a new proposal.
+    Tests that a user can generate a new proposal and records the video.
     """
+    # 1. Setup constants
     email = "test_user@unhcr.org"
     password = "password123"
     base_url = "http://localhost:8502"
-
-    page.goto(f"{base_url}/login")
-    page.get_by_test_id("email-input").fill(email)
-    page.get_by_test_id("password-input").fill(password)
-    page.get_by_test_id("submit-button").click()
-
-    expect(page).to_have_url(re.compile(".*dashboard"))
     
-    page.get_by_test_id("new-proposal-button").click()
-    
-    expect(page).to_have_url(re.compile(".*chat"))
+    # Define where the video will be saved.
+    VIDEO_DIR = "playwright/test-results/videos"
 
-   # Take screenshot 
-    page.screenshot(path="playwright/test-results/4_generate_proposal.png")
-
-
-    # Fill in the proposal form
-    page.get_by_role("textbox", name="Project Draft Short name*").fill('Refugee Children Education Initiative')
-    page.get_by_placeholder('Provide as much details as possible on your initial project idea!').fill('Establishing a comprehensive primary education program for 2,500 refugee children aged 6-14.')
-
-    # Main Outcome (multiselect)
-    page.locator("label").filter(has_text="Main Outcome*").click()
-    page.locator(".main-outcome__input-container").click()
-    page.get_by_role("option", name="OA11-Education").click()
-    page.locator(".main-outcome__input-container").click()
-    page.get_by_role("option", name="OA7-Community").click()
-
-    page.get_by_test_id("beneficiaries-profile").click()
-    page.get_by_test_id("beneficiaries-profile").fill("2,500 refugee children aged 6-14")
-    
-    page.get_by_test_id("potential-implementing-partner").click()
-    page.get_by_test_id("potential-implementing-partner").fill("UNHCR, UNICEF, Save the Children")
-
-    # Geographical Scope (select)  -------
-    page.get_by_label('Geographical Scope').select_option(label='One Country Operation')
-
-    # Country / Location(s) (creatable select)  
-    page.locator(".country-location-s__input-container").click()
-    page.get_by_role("option", name="Colombia").click()
+    # Ensure the directory exists
+    os.makedirs(VIDEO_DIR, exist_ok=True)
 
 
-    # Budget Range (creatable select) -------
-    page.locator(".budget-range__input-container").click()
-    page.get_by_role("option", name="1M$").click()
+    # Use the sync_playwright context manager to launch and control the browser lifecycle
+    with sync_playwright() as playwright:
+        # Launch browser (use chromium, firefox, or webkit)
+        browser = playwright.chromium.launch(headless=False, slow_mo=500)
+        
+        # 2. Create a new context and set the video recording directory
+        # Video recording starts now.
+        context = browser.new_context(
+            record_video_dir=VIDEO_DIR,
+            # Set viewport to a high resolution (e.g., Full HD) for maximum screen space
+            viewport={"width": 1920, "height": 1080}, 
+            # Set the video output size to match the viewport for best quality
+            record_video_size={"width": 1920, "height": 1080}
+        )
+        
+        # 3. Get a new page from the context
+        page = context.new_page()
 
-    # Duration (creatable select)  -------
-    page.locator(".duration__input-container").click()
-    page.get_by_role("option", name="12 months").click()
+        # -------------------
+        # Start of Test Logic
+        # -------------------
+        page.goto(f"{base_url}/login")
+        page.get_by_test_id("email-input").click()
+        page.get_by_test_id("email-input").fill(email)
+        page.get_by_test_id("password-input").click()
+        page.get_by_test_id("password-input").fill(password)
+        page.get_by_test_id("submit-button").click()
+        expect(page).to_have_url(re.compile(".*dashboard"))        
+        page.get_by_test_id("new-proposal-button").click()        
+        expect(page).to_have_url(re.compile(".*chat"))
 
-    # Targeted Donor (creatable select)  -------
-    page.locator(".targeted-donor__input-container").click()
-    page.get_by_role("option", name="Sweden - Ministry for Foreign").click()
-    
-    # Click the "Generate" button  -------
-    page.get_by_role('button', name='Generate').click()
+        # Fill in the proposal form ------- 
+        page.get_by_test_id("project-draft-short-name").click()
+        page.get_by_test_id("project-draft-short-name").fill('Project: Refugee Children Education Initiative')
+        page.get_by_role("textbox", name="Provide as much details as").click()
+        page.get_by_placeholder('Provide as much details as possible on your initial project idea!').fill('Establishing a comprehensive primary education program for 2,500 refugee children aged 6-14.')
 
-    # Wait for the sections to be generated. This can be slow.  -------
-    expect(page.get_by_role("button", name="Download Document")).to_be_visible(timeout=120000)
-   # Take screenshot  
-#    page.screenshot(path="playwright/test-results/5_proposal_created.png")
+        # Main Outcome (multiselect) ------- 
+        page.locator(".main-outcome__input-container").click()
+        page.get_by_role("combobox", name="Main Outcome").fill("ed")
+        page.get_by_role("option", name="OA11. Education").click()
+        page.get_by_role("combobox", name="Main Outcome").fill("com")
+        page.get_by_role("option", name="OA7. Community Engagement and").click()
 
-#        with page.expect_download() as download_info:
- #       page.get_by_role("button", name="Download Document").click()
- #   download = download_info.value
- #   page.get_by_role("button", name="Peer Review").click()
-   # page.get_by_test_id("user-select-checkbox-be7ebe30-fcc8-442f-993b-dd89b8d61dfb").check()
-    #page.get_by_test_id("deadline-input").fill("2025-09-10")
-    #page.get_by_test_id("confirm-button").click()
+        # Beneficiaries ------- 
+        page.get_by_test_id("beneficiaries-profile").click()
+        page.get_by_test_id("beneficiaries-profile").fill("2,500 refugee children aged 6-14")
+        
+        # Partner ------- 
+        page.get_by_test_id("potential-implementing-partner").click()
+        page.get_by_test_id("potential-implementing-partner").fill("UNHCR, UNICEF, Save the Children")
+
+        # Geographical Scope (select)  ------- 
+        page.get_by_test_id("geographical-scope").select_option("One Country Operation")
+      
+        # Country / Location(s) (creatable select)  
+        page.locator(".country-location-s__input-container").click()
+        #page.get_by_role("option", name="Colombia").click()
+        page.get_by_role("option", name="Afghanistan").click()    
+
+        # Budget Range (creatable select) -------
+        page.locator(".budget-range__input-container").click()
+        page.get_by_role("option", name="1M$").click()
+
+        # Duration (creatable select)  -------
+        page.locator(".duration__input-container").click()
+        page.get_by_role("option", name="12 months").click()
+
+        # Targeted Donor (creatable select)  -------
+        page.locator(".targeted-donor__input-container").click()
+        page.get_by_role("option", name="Sweden - Ministry for Foreign").click()
+
+
+        # Optional Screenshot - Video is being recorded anyway
+        page.screenshot(path="playwright/test-results/4_generate_proposal.png")
+        
+        # Click the "Generate" button  -------
+        page.get_by_role('button', name='Generate').click()
+
+        # Wait for the sections to be generated. This can be slow.  -------
+
+        # Wait for the sections to be generated. This can be slow.  -------
+        # The wait is now conditioned on the visibility of the edit button for the first section.
+        expect(page.get_by_test_id("section-options-0").get_by_role("button", name="edit-section-")).to_be_visible(timeout=500000)
+        
+
+#    page.get_by_role("heading", name="Project: Refugee Children").first.click()
+        page.get_by_role("complementary").get_by_text("Monitoring").click()
+        page.get_by_test_id("section-options-4").get_by_role("button", name="edit-section-").click()
+        page.get_by_role("button", name="edit-section-").click()
+        page.screenshot(path="playwright/test-results/6_edit_section.png")
+        page.get_by_text("Proposal Prompt").click()
+    # page.get_by_role("button", name="Button Icon Manage Knowledge").click()
+    # page.get_by_role("button", name="Cancel").click()
+    # page.get_by_role("button", name="Peer Review").click()
+    # page.get_by_test_id("user-select-checkbox-5c092577-1230-4473-acf8-b6e3bfb02bca").check()
+    # page.get_by_test_id("deadline-input").fill("2025-11-19")
+    # page.get_by_test_id("confirm-button").click()
+    # page.get_by_test_id("logo").click()
+
+
+        #page.get_by_role("button", name="Regenerate").click()
+        #page.locator("#regenerate-prompt").fill("revise this section to fit in 200 characters")
+        #page.get_by_role("button", name="Button Icon Regenerate").click()
+        #page.get_by_role("button", name="edit-section-").click()
+        #page.get_by_test_id("section-options-0").get_by_role("button", name="edit-section-").click()
+        #page.get_by_role("textbox", name="editor for Summary").click()
+        #page.get_by_role("textbox", name="editor for Summary").press("Revising this...")
+
+        page.get_by_role("button", name="Download Document").click()
 
 
 
-    page.get_by_role("button", name="Peer Review").click()
-    page.get_by_text("Test User bis").click()
-    page.get_by_test_id("deadline-input").fill("2025-09-11")
-    # Take screenshot  
-    page.screenshot(path="playwright/test-results/6_peer_review.png")
-    page.get_by_test_id("confirm-button").click()
+        page.get_by_test_id("logo").click()
+        #page.get_by_role("article").filter(has_text="Refugee Children Education Initiative Establishing a comprehensive #primary").get_by_test_id("project-options-button").click()
 
-    page.get_by_test_id("logo").click()
-    page.get_by_role("article").filter(has_text="Refugee Children Education Initiative Establishing a comprehensive primary").get_by_test_id("project-options-button").click()
+        # -------------------
+        # End of Test Logic
+        # -------------------
 
-    #expect(page.locator('h2:has-text("Background and Needs Assessment")')).to_be_visible(timeout=120000)
+        # 4. Close the context and browser
+        # The video file is saved when the context closes.
+        video_path = page.video.path()
+        context.close()
+        browser.close()
+        
+        # Optional: Rename the file to something more descriptive
+        new_video_path = os.path.join(VIDEO_DIR, "proposal_generation.webm")
+        os.rename(video_path, new_video_path)
+        print(f"Video saved successfully to: {new_video_path}")
 
-   # browser.close()
